requirements.txt
# Windmill ML Pipeline Dependencies
pika>=1.3.0           # RabbitMQ message queue client
psycopg2>=2.9.0; sys_platform == "darwin"           # PostgreSQL client (Mac source build)
psycopg2-binary>=2.9.0; sys_platform != "darwin"      # PostgreSQL client (Linux binary)
python-dotenv>=1.0.0  # Environment variable loading
requests>=2.31.0      # HTTP client for ML service calls
PyYAML>=6.0           # YAML configuration file parsing
Pillow>=10.0.0        # PIL/Pillow for image manipulation and cropping
pillow-heif>=0.16.0   # HEIC/HEIF image format support (iPhone photos)
Flask>=3.0.0          # HTTP API for single image submission
nltk>=3.8.0           # WordNet synonym collapsing for noun consensus
spacy>=3.8.0          # Noun extraction from VLM captions (en_core_web_sm model)
imagehash>=4.3.0      # Perceptual hashing for image deduplication across formats/resolutions
anthropic>=0.50.0    # Anthropic SDK — used by caption_summary_worker for Claude synthesis
